requirements.txt
fastapi==0.109.0
requests==2.31.0
pydantic==2.5.3
uvicorn==0.25.0
google-generativeai==0.5.2
python-dotenv==1.0.1
